requirements.txt
appnope==0.1.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
backports.shutil-get-terminal-size==1.0.0
blessings==1.7
bpython==0.17.1
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.1
curtsies==0.3.0
decorator==4.1.2
Django==2.2.0
djangorestframework==3.10.1
enum34==1.1.6
factory-boy==2.10.0
Faker==0.8.12
greenlet==0.4.15
idna==2.8
importlib-metadata==0.18
ipaddress==1.0.19
ipython==5.5.0
ipython-genutils==0.2.0
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.0
parameterized==0.6.1
pathlib2==2.3.0
pexpect==4.3.1
pickleshare==0.7.4
pluggy==0.12.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.8.0
Pygments==2.2.0
pylint==2.3.1
pyparsing==2.4.0
pytest==5.0.1
pytest-django==3.5.1
python-dateutil==2.7.2
pytz==2017.3
requests==2.21.0
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
South==1.0
text-unidecode==1.2
tox==3.0.0
traitlets==4.3.2
typed-ast==1.4.0
typing==3.6.6
urllib3==1.24.1
virtualenv==16.6.2
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.5.2
